set endpoint logic
